Splitting RBAC test to smaller pieces.
